Removed maxAbi
